add some exceptional handling

+++ unmount.bash
@@ -5,7 +5,15 @@ then
     exit -1
 fi
 sudo umount /media/yrk16/$1
-sudo rm -r /media/yrk16/$1
+sudo test -d /media/yrk16/$1
+if [ 0 -eq $? ]
+ then
+    sudo rm -r /media/yrk16/$1
+fi
 sudo umount ./$1
-sudo rm -r ./$1
+sudo test -d ./$1
+if [ 0 -eq $? ]
+ then
+    sudo rm -r ./$1
+fi
 exit
